Default unit to hours - PMT #104842
This is up for debate, let me know what you think about this - Maurice
thinks the time unit should default to hours. Currently, entering a
simpleduration with no unit into PMT is a bit broken.. it creates the
time object and always gives it a duration of 0. This change would fix
that behavior, but we could also solve it by telling the user they need
to enter a time unit when they don't include one.